- FUNCTIONALITY CHANGE: Previously, "tables" were text files holding data which was strictly two-dimensional.  When using "unu lut" and "unu rmap", its often very conveniant to put the lut/map in a text file, but these are 1D, not 2D.  So, now tables can represent 1D data.  This has been implemented by allowing
  # dimension: 1

to be a "comments" preceeding the table data.  The default
dimension for tables is still 2, the line above will
explicitly set it to 1.  But the dimension can not be set to
anything other than 1 and 2, because these are the only
dimensions in which we can imply the axis sizes by the text
layout of the numbers.  Also, "# dimension: 1" can not
appear in the comments *after* a per-axis field, such as
"# spacings: 1 2".


git-svn-id: file:///home/johnsonhj/src/TEEM_TRANSITION/teem_svn2git/local_svn_repo_mirror/code/teem/trunk@625 3d70eeeb-363e-0410-a505-8a46323a89f2