added two new fields to the limnCamera, "fov" and "aspect", augmented the semantics of limnCameraUpdate, and made an effort to fix/debug the limnCameraPathMake() function.
"fov" is like the "fovy" argument to gluPerspective()- the angle in
degrees vertically subtended by the view window, and "aspect" is the
ratio of horizontal to vertical size of the view window.

Now, when limCameraUpdate() is called, it will see if both fov and aspect
are set (non-NaN), and if so will update uRange[] and vRange[] accordingly.
Otherwise, uRange[] and vRange[] will not be touched.

limnCameraPathMake is not a new function, but it never worked usefully.


git-svn-id: file:///home/johnsonhj/src/TEEM_TRANSITION/teem_svn2git/local_svn_repo_mirror/code/teem/trunk@2180 3d70eeeb-363e-0410-a505-8a46323a89f2