Some restructing:
* All the limnContour3DContext-related code has been moved to a
new library, called "seek", for general-purpose polygonal feature
extraction

* In limnPolyData, there are now per-attribute "Num" variables,
insteadof just a single vertNum; this was needed to clarify the logic
of how per-attribute arrays were allocated and maintained.

* limn now depends on gage.  At one point (prior to this check-in),
this was because of how the feaure extraction needed gage, but now it
is simply because there is at the plan of having more gage-specific
per-vertex attribute information stored in the limnPolyData.


git-svn-id: file:///home/johnsonhj/src/TEEM_TRANSITION/teem_svn2git/local_svn_repo_mirror/code/teem/trunk@3309 3d70eeeb-363e-0410-a505-8a46323a89f2