Various changes in anisotropy calculation: Added the _tenAnisoTen_f[] and _tenAnisoTen_d[] arrays of functions, and the associated tenAnisoTen_f() and tenAnisoTen_d() function wrappers, to complement tenAnisoEval_{f,d}(): these provide a way of evaluating a given anisotropy measure given the tensor, whereas tenAnisoEval_{f,d}() provide a way of evaluating a given aniosotropy given the soted eigenvalues.
Removed the tenAniso_Cz anisotropy measure.  It was never used in Teem,
its not used in the literature, and its odd because its unbounded.

Renamed tenInvariantKGradients_d --> tenInvariantGradientsK_d
Renamed tenInvariantRGradients_d --> tenInvariantGradientsR_d
(this is in keeping with Teem convention of differentiating similar
symbols as late as possible in the name)


git-svn-id: file:///home/johnsonhj/src/TEEM_TRANSITION/teem_svn2git/local_svn_repo_mirror/code/teem/trunk@3539 3d70eeeb-363e-0410-a505-8a46323a89f2