Some API additions, an API change (for a function nobody should have been using), and tweaks to string-based kernel identification
API NEW: added nrrdSaveMulti(), which is like nrrdSave(), but saves
an array of Nrrds to some set of files, based on a sprintf-style
formatted string.

API CHANGE: changed:
  _nrrdContainsPercentDAndMore(const char *str)
to the slightly more general purpose:
  _nrrdContainsPercentThisAndMore(const char *str, char this) 
The only reason this was a public function was so that unrrdu/dice
could use it, but unrrdu/dice should be changed to use the new
nrrdSaveMulti anyway.

API NEW: added

  int nrrdKernelSpecSprint(char str[AIR_STRLEN_LARGE],
                           NrrdKernelSpec *ksp)

which is like the opposite of nrrdKernelSpecParse(), it prints the
kernel specification to a string.  New code added to test/tkernel.c
to demo this (as well as adding airMop use).

To make the sprinted result parse'able, it was discovered that
sometimes the kernel's self-identification wasn't actually recognized
as identifying that kernel by nrrdKernelParse().  This has been fixed.

As part of that clean up, some single-letter kernel identifications
were removed from _nrrdKernelStrToKern, because they were too clever,
and too apt to be confused:

-  if (!strcmp("z", str))          return nrrdKernelZero;
-  if (!strcmp("b", str))          return nrrdKernelBox;
-  if (!strcmp("t", str))          return nrrdKernelTent;
-  if (!strcmp("c", str))          return nrrdKernelBCCubic;
-  if (!strcmp("cd", str))         return nrrdKernelBCCubicD;
-  if (!strcmp("cdd", str))        return nrrdKernelBCCubicDD;
-  if (!strcmp("q", str))          return nrrdKernelAQuartic;  
-  if (!strcmp("qd", str))         return nrrdKernelAQuarticD;  
-  if (!strcmp("qdd", str))        return nrrdKernelAQuarticDD;  
-  if (!strcmp("g", str))          return nrrdKernelGaussian;



git-svn-id: file:///home/johnsonhj/src/TEEM_TRANSITION/teem_svn2git/local_svn_repo_mirror/code/teem/trunk@3863 3d70eeeb-363e-0410-a505-8a46323a89f2