A flurry of activity leading to two things:
1) a change in the API of nrrdCompare, nrrdAxisInfoCompare, and
nrrdArrayCompare: it was too annoying to have their return be
analogous to strcmp, while all error handling had to be via setting of
flags passed by pointer.  Now, they always use biff in case of actual
error, and the difference (as in the return from strcmp) is saved into
an int*.  This includes removing the short-lived and ill-advised
_NRRD_ERROR_RETURN. There are recently added anyway so its not an API
CHANGE.

2) fixing some long-standing bugs with NRRD format IO, and clarifying
what needs to be looked at in the future.  A lengthy new comment at
the top of nrrd/formatNRRD.c documents the state of things.
Basically, the file format and the library spec could both stand to be
more explicit about what characters are allowed in what strings
(things like labels, units, key/value pairs, content, comments), and
if they are allowed, how they are to be escaped or encoded on writing.
Certainly something that merits more time and attention in the long
term.
An example of the prior bug: a key or value in a key/value pair can
be added that contains \n, but writing this would have created a 
broken NRRD file.  Similar issues existed for labels, units, and
space units.  So the NRRD writer is now more vigilant about not
creating a broken NRRD file.  A new nrrd/keyvalue.c/_nrrdWriteEscaped
is one of the outcomes of this, but now its utility is beyond
just dealing with key/value pairs.
Another prior bug, mysteriously (to GLK) undetected by const-correctness:
airOneLinify was being called on nrrd->content and nrrd->sampleUnits 
on *write*.


git-svn-id: file:///home/johnsonhj/src/TEEM_TRANSITION/teem_svn2git/local_svn_repo_mirror/code/teem/trunk@5364 3d70eeeb-363e-0410-a505-8a46323a89f2